Extracted processValidatorsEvents to commons
